requirements.txt
anthropic==0.41.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==1.13.0
click==8.1.7
rich==13.7.0
schedule==1.2.0
python-dateutil==2.8.2
